pyproject.toml
[project]
name = "catat"
version = "0.1.0"
description = "This project is the backend layer for catat"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=43.0.3",
    "fastapi[standard]>=0.115.4",
    "httpx>=0.27.2",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.9.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.0.1",
    "sqlmodel>=0.0.22",
]
